a/b test unit tests and cleanups
